Add csvHelper converter

+++ back/Billing/Billing.Contracts/Billing.Contracts.Domain/Offers/Parsing/OfferRow.cs
@@ -15,8 +15,8 @@ namespace Billing.Contracts.Domain.Offers.Parsing
         public DateTime? StartsOn { get; set; }
         public int MinIncludedCount { get; set; }
         public int MaxIncludedCount { get; set; }
-        public double UnitPrice { get; set; }
-        public double FixedPrice { get; set; }
+        public decimal UnitPrice { get; set; }
+        public decimal FixedPrice { get; set; }
     }
